Refactor clasp setting test
